feat: Add loading state to useEntryGuard and separate verification logic
- Added a  state to  to display a loading indicator during async operations.
- Improved navigation flow by showing a spinner and Loading... text while fetching user and room data.
- Handled errors gracefully by stopping the loading state when an error occurs.
- Enhanced user experience with smoother page transitions during redirects.
- Separated email verification logic into its own  file for better modularity and maintainability.
