+++ frontend/src/app/(auth)/signup.tsx
@@ -1,35 +1,17 @@
 import * as React from 'react';
-import { useEffect, useState } from 'react';
-import { Text, TextInput, View } from 'react-native';
+import { useState } from 'react';
+import { Text, View } from 'react-native';
 import { useSignUp } from '@clerk/clerk-expo';
 import Button from '@/components/Button';
 import { useRouter } from 'expo-router';
 import FormInput from '@/components/FormInput';
-import { createUser } from '@/apis/user';
 
 export default function SignUpForm() {
-  const { isLoaded, signUp, setActive } = useSignUp();
+  const { isLoaded, signUp } = useSignUp();
   const router = useRouter();
   const [emailAddress, setEmailAddress] = useState('');
   const [password, setPassword] = useState('');
-  const [pendingVerification, setPendingVerification] = useState(false);
-  const [code, setCode] = useState('');
   const [error, setError] = useState<string | null>(null);
-  const [resendTimer, setResendTimer] = useState(0);
-
-  // This useEffect creates a real-time countdown for the resend button.
-  useEffect(() => {
-    let interval: any;
-
-    if (resendTimer > 0) {
-      interval = setInterval(() => {
-        setResendTimer((prev) => prev - 1);
-      }, 1000); // every 1 second (1000), minus 1 ((prev) => prev - 1)
-    }
-
-    return () => clearInterval(interval);
-    // clearInterval delete the interval when the value of resendTimer changes
-  }, [resendTimer]);
 
   // ✅ Sign up with email & password
   const onSignUpPress = async () => {
@@ -37,12 +19,20 @@ export default function SignUpForm() {
     console.log('📩 Sign-up input:', { emailAddress, password });
 
     try {
+      // Create a new user with Clerk
       await signUp.create({
         emailAddress,
         password,
       });
+
+      // Prepare email verification
       await signUp.prepareEmailAddressVerification({ strategy: 'email_code' });
-      setPendingVerification(true);
+
+      // Navigate to the verify-email page
+      router.replace({
+        pathname: '/(auth)/verify-email',
+        params: { emailAddress }, // Pass emailAddress to the verification page
+      });
     } catch (err: any) {
       console.error('Sign-up error:', JSON.stringify(err, null, 2));
       const readableMessage =
@@ -54,126 +44,6 @@ export default function SignUpForm() {
     }
   };
 
-  // ✅ Verify code
-  const onVerifyPress = async () => {
-    if (!isLoaded) return;
-
-    try {
-      // 1) Attempt verification & grab the new user’s ID
-      const { status, createdSessionId, createdUserId } =
-        await signUp.attemptEmailAddressVerification({ code });
-
-      // 2) Bail if verification didn’t complete
-      if (status !== 'complete' || !createdUserId) {
-        setError('Verification failed. Please try again.');
-        return;
-      }
-
-      // 3) Activate the new session
-      await setActive({ session: createdSessionId });
-
-      // 4) Save user to your DB
-      const createdUser = await createUser({
-        email: emailAddress,
-        user_id: createdUserId,
-      });
-      console.log('✅ New user saved to database:', createdUser);
-
-      // 5) Redirect
-      router.replace({
-        pathname: '/(onboard)/onboarding-flow',
-        params: { user_id: createdUserId }, // Pass user_id here
-      });
-    } catch (err: any) {
-      console.error('🚨 onVerifyPress error:', err);
-      const msg =
-        err?.errors?.[0]?.shortMessage ||
-        err?.errors?.[0]?.longMessage ||
-        err?.message ||
-        'Something went wrong during verification. Please try again.';
-      setError(msg);
-    }
-  };
-
-  // ✅ Resend verification code
-  const onResendCode = async () => {
-    if (!isLoaded) return; // Make sure Clerk is fully loaded before trying to resend the code.
-    try {
-      await signUp.prepareEmailAddressVerification({ strategy: 'email_code' }); // Resend the verification code
-      setResendTimer(120); // set 2 minutes timer
-    } catch (err: any) {
-      console.error('Resend code error:', JSON.stringify(err, null, 2));
-      const readableMessage =
-        err?.errors?.[0]?.shortMessage ||
-        err?.errors?.[0]?.longMessage ||
-        err?.message ||
-        'Failed to resend verification code. Please try again.';
-      setError(readableMessage);
-    }
-  };
-
-  // ✅ Sign up with Google (OAuth)
-  const onGoogleSignUpPress = async () => {
-    //need help here
-  };
-
-  // ✅ OTP Verification UI
-  if (pendingVerification) {
-    return (
-      <View>
-        <View className="w-[300px]">
-          <Text className="text-lg mb-4 text-accent" style={{ fontFamily: 'Poppins-Bold' }}>
-            Verify your email
-          </Text>
-          <TextInput
-            value={code}
-            placeholder="Enter your verification code"
-            onChangeText={setCode}
-            className="border border-accent bg-white px-4 py-3 rounded-lg w-[300px] mb-4"
-          />
-
-          {/* Show error if exists */}
-          {error ? (
-            <Text
-              className="text-red-600 text-center mb-2 w-[300px]"
-              style={{ fontFamily: 'Poppins-Regular' }}
-            >
-              {error}
-            </Text>
-          ) : null}
-
-          <Button
-            label="Verify"
-            onPress={onVerifyPress}
-            size="py-3 px-4"
-            color="bg-accent"
-            className="w-[300px] items-center"
-            textClassName="text-white text-[16px]"
-            textStyle={{ fontFamily: 'Poppins-Regular' }}
-          />
-
-          {/* ✅ Resend button or Timer */}
-          {resendTimer > 0 ? (
-            <Text className="text-accent mb-4" style={{ fontFamily: 'Poppins-Regular' }}>
-              You can resend code in {resendTimer} seconds
-            </Text>
-          ) : (
-            <Button
-              label="Resend Code"
-              onPress={onResendCode}
-              size="py-3 px-4"
-              color="border border-accent"
-              className="w-[300px] items-center mb-4"
-              textClassName="text-accent text-[16px]"
-              textStyle={{ fontFamily: 'Poppins-Regular' }}
-            />
-          )}
-        </View>
-      </View>
-    );
-  }
-
-  // ✅ Sign-up form UI
   return (
     <View>
       <View className="w-[300px]">
@@ -182,7 +52,7 @@ export default function SignUpForm() {
           borderColor="#F5829B"
           autoCapitalize="none"
           value={emailAddress}
-          placeholder="Sample@domain.com"
+          placeholder=""
           onChangeText={setEmailAddress}
         />
         <FormInput
@@ -190,7 +60,7 @@ export default function SignUpForm() {
           borderColor="#F5829B"
           autoCapitalize="none"
           value={password}
-          placeholder="•••••••••••"
+          placeholder=""
           secureTextEntry={true}
           onChangeText={setPassword}
         />
@@ -219,7 +89,7 @@ export default function SignUpForm() {
       {/* Button to Sign Up with Google */}
       <Button
         label="Sign up with Google"
-        onPress={onGoogleSignUpPress}
+        onPress={() => console.log('Google Sign Up')} // Placeholder for Google sign-up logic
         size="py-3 px-4"
         color="border border-accent"
         className="w-[300px] items-center mb-3"
